pyproject.toml
[tool.poetry]
name = "fcs"
version = "0.1.0"
description = ""
authors = ["ricepotato <sukjun40@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.16"
fastapi = "^0.89.1"
pydantic = "^1.10.4"
boto3 = "^1.26.60"
python-dotenv = "^0.21.1"
uvicorn = "^0.20.0"
mypy-boto3-dynamodb = "^1.26.24"
pymongo = "^4.3.3"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"
pytest = "^7.2.1"
mypy = "^0.991"
pylint = "^2.15.10"
botocore-stubs = "^1.29.62"
boto3-stubs = "^1.26.62"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
